requirements.txt
python-decouple
jinja2
mysql-connector-python